CSV to NPY conversion
I got the csv to npy conversion to work. Run the file in the folder csv2npy. Currently, it only supports a 1D array, so prefilter the data you have downloaded to only show the Adjusted Closing value of each stock and save that into a csv.

Also, there is an error where the main trading_agent program doesn't run anymore because of a new error. I included a picture of the error I got. We should probably fix that next.