fixed csv2npy plus other changes
i fixed csv2npy so that it ignores the first line of data (the header). i changed some other configuation stuff in dqn_agent, too. mostly to see if we can start feeding the data back into the machine, so that it can actually learn from its horrible mistakes and one day make me millions of dollars buying and selling stocks on the stock market. Or at least get me a good grade on this project. I'll take either.